feat: implement a WDL lexer using `logos`.
This commit introduces an `experimental` module as a placeholder for a new
parser implementation.

To start, a lexer implementation based on `logos` has been added.

The lexer may have several different token types depending on context, mainly
to support interpolation in string literals and commands.

The implementation should be enough to lex a 1.x WDL document.

The parser implementation that will make use of the lexer is forthcoming.