feat: implement a new experimental concrete syntax tree type.
This commit adds a new concrete syntax tree type called `SyntaxTree` that is
based on `rowan` for the underlying representation.

The tree is constructed via an infallible `parse` method that always returns a
`SyntaxTree` along with a list of parse errors.

A `Parser` has also been added that takes WDL source and outputs a list of
parser events that can be used to build a `SyntaxTree`.

Currently only version statements are being parsed in the V1 grammar; more
grammar will be incrementally introduced in the future.

Additionally, new file-based tests have been added to the `wdl-grammar` crate;
each test has a `source.wdl` input and a baseline CST and error list files to
compare against. The baselines are easily updated by setting the `BLESS`
environment variable and running the test.

All of the above if feature-gated behind the `experimental` feature of
`wdl-grammar`.