feat: implement parsing `meta` and `parameter_meta` in experimental parser.
This commit implements parsing of `meta` and `parameter_meta` sections in both
tasks and workflows for the experimental parser.

This also includes a fix for lexing single-digit integers, a fix for a panic when
parsing a `Map` key type that is not a primitive type name, and fixes to the
`delimited` and `expect*` methods of `Parser`.

Error messages for unmatched braces and brackets have been introduced as well.

Unknown tokens are now properly added to the syntax tree too.

This commit also moves the `Event::Error` variant out into its own collection.

Doing so reduces the `Event` size from 48 bytes to 24 bytes; as there will be an
event for every token in the file, the size of `Event` is important.

Errors are now tracked directly in the parser implementation and returned when
the parser is finished.