feat: implement input and output sections in the experimental parser.
This commit implement parsing of input and output sections in the experimental
parser for both tasks and workflows.

Also implemented is the caching of a lexer peek operation which prevents the
lexer from having to tokenize the input again.

The `expect*` methods of `Parser` now peek and move next only if the token
matches what's expected; this will prevent consuming recovery tokens in
recovery contexts.