fix: fix placeholder option parsing in the experimental parser.
This commit fixes the parsing of placeholder options by adding a `peek2` method
to check for an assignment token following the option identifier.

Without the assignment token present, it should treat the contents as an
expression, such as a call to the `sep` function.

Currently `Lexer::peek` is retokenizing on every `peek` followed by `next`; a
forthcoming PR will properly cache the result of `peek` so that re-tokenization
will only ever occur for `peek2`, which will only be called for string
placeholders.