fix: allow reserved names in the experimental parser.
The experimental parser is too strict regarding reserved names in the parse,
specifically in metadata and parameter metadata sections, where keys are
allowed to be reserved names.

This commit fixes the parser to accept reserved names in addition to
identifiers; the places where reserved names are not allowed will be checked
during (forthcoming) validation of the tree instead.

This also fixes the incorrect recovery in metadata sections which might cause
an infinite loop: the recovery was incorrectly set to recovery at a value
instead of any identifier, which meant that recovery would move to a token that
would immediately fail, recovery would then not move, and the cycle would
repeat.

Added tests for both problems.

Additionally the `NameNode` and `QualifiedNameNode` nodes have been removed
from the tree in favor of simply looking for "any identifier token child" in
the forthcoming AST.