fix: recover past interpolations in the experimental parser.
This commit implements recovering past interpolations in the experimental
parser; if recovery encounters the start of an interpolation, it delegates to
the token type to move the parser past the interpolation.

This also fixes recovery in brace commands, where it would not consume the
closing placeholder brace, causing a premature ending to the command. Likewise,
for the other interpolations, fixes recovery to also consume the closing
placeholder brace so it does not appear in the literal text.

Removes the "metadata string cannot contain a placeholder" error in favor of
just treating them as literal text (there's at least one source file in
gauntlet that has them, expecting the literal text).