fix: fix disambiguation of struct literals in the experimental parser.
This commit fixes the disambiguation of struct literal expressions in the
experimental parser.

Previously, we looked ahead from the identifier by one token to see if it is an
open brace; if so, we would assume it is a struct literal.

However, this conflicts with other grammar rules that expect an open brace
following an expression, such as conditional statements in workflows.

The fix is to look ahead three tokens for `{`, <ident>, and `:` to consider it
a struct literal; if those tokens aren't present, then the expression will be
treated as a name reference.

Note that this means that empty struct literals (and by extension struct
definitions themselves) are invalid; while the WDL spec doesn't clarify this
requirement, the example WDL parsers from OpenWDL do require at least one
member specifier in a struct literal as well.