feat: AST validation for the experimental parser.
This commit adds validation to the new AST representation.

Validation is performed via `Validator`; the validation is performed via a set
of `Visitor` implementing types. The AST is walked once and the visitors are
called depending on the nodes and tokens that are visited.

The following validation rules are implemented:

* Checking for duplicate workflows, sections, missing command sections, and
  empty structs.
* Checking for duplicate keys in runtime and metadata sections, as well as
  object and struct literals.
* Checking out-of-bounds numerical literals.
* Checking for invalid escape sequences in strings or for characters that must
  be escaped.

In addition, the error format of the `wdl-grammar` file tests has change from
`miette` to use `codespan_reporting`, which is what `sprocket` uses to format
diagnostics.