feat: port preamble comment and whitespace lint rules.
This commit ports the preamble comment and whitespace lint rules to the new
parser.

It adds supports for visiting whitespace and comment tokens as well.

Token visitations no longer get called twice as they never have children, so
the `VisitReason` was removed as a parameter.

Also contains a fix to the comment tokenization so that comments don't capture
a carriage return on `\r\n` line endings.