feat: migrate `wdl-gauntlet` to the new parser implementation.
This commit removes the use of the existing parser implementation in favor of
the new implementation in `wdl-gauntlet`.

As a result, `Arena.toml` and `Gauntlet.toml` have been refreshed; in the case
of `Arena.toml`, there's a ~10X increase in logged diagnostics due to the new
parser implementation successfully parsing files that the existing parser
implementation incorrectly parsed with errors.

Also added some wall-clock timing for the time spent analyzing a source file
and an aggregate duration displayed at the end of gauntlet's output. From the
before and after comparisons, the new parser implementation appears to be at
least 20X faster in analyzing files, likely due to using `logos` for lexing and
the new parser implementation does not backtrack like `pest` does.

Also added configuration for ignoring specific files so that
`task-templates.wdl` can be excluded as it isn't expected to be valid WDL (it
has non-WDL placeholders in it to serve as a template).

Once this merges and we're in consensus, removal of the existing parser
implementation can begin.