fix: add a label for suggesting where to put a missing delimiter.
This commit changes the `delimited` method of `Parser` to add a label to the
expected diagnostic for a missing delimiter.

It does this by finding the last non-trivia token event and using its span to
point at where the missing delimiter might go.

Also fixes a failure to consume a delimiter if recovery stops at a delimiter.

Fixes #67.