fix: make trivia locations in the CST consistent.
This fixes where trivia is located in the CST so that it is consistent; the
previous inconsistency was the result of attaching trivia to the CST whenever a
`peek` operation was performed.

Now any trivia encountered during a `peek` is buffered and attached only when
we consume the next token from the stream or before a node is started.

This also fixes sorting the diagnostics returned from parsing and validation
such that it is sorted by primary label offset.

Fixes #81.