fix: improve diagnostics that have a label pointing at a string.
This fix improves any diagnostic that has a label pointing at a string.

Previously the label would point just at the opening quote of the string
because we don't know the span of the string until after we recover; the
recovery asks the lexer to move past any interpolations it encounters.

The fix is to pass the diagnostic we're recovering for to the `recover` method
and adjust any labels that start an interpolation context to include the any
source that was moved past.

Fixes #80.