fix: improve diagnostics for strings used as member names in struct literals.
This commit improves the diagnostics around encountering a struct literal that
specifies the member names as strings instead of the required identifier.

The problem is that strings are interpolated, so the parser cannot "see" the
colon following the string without recovering past the interpolation.

The fix is to have `literal_struct_or_name_ref` check for the second token
peeked to be the start of a string, which is not legal syntax in as expression.
If the second token does start a string, recover past the string and check to
see if the next token would be a colon. If it is a colon, treat it as a literal
struct expression; otherwise, treat it as a name reference expression.

Fixes #78.