refactor: make `Visitor` and implementations version agnostic.
Currently, AST visitation uses `SyntaxKind` to determine which methods on a
visitor to call. As `SyntaxKind` is intentionally the union of all possible WDL
versions, there's no reason for the `Visitor` trait to be in a `v1` module.

This commit moves the `Visitor` trait to the root of the crate and also moves
any implementations of `Visitor` out of a `v1` module as well. For validation,
they now go in a `validation` module. For lint rules, they go in a `rules`
module.

For the lint rules, it means that if a lint rule works only on certain nodes or
tokens (e.g. whitepsace) that aren't tied to a particular version of WDL or the
syntax they operate on doesn't change from version to version, the rule will
continue to operate on newer versions of WDL.

As new syntax is introduced over time, rules that do need to operate on the
newer syntax will need to implement additional `Visitor` trait methods to visit
the new syntax nodes.

If for whatever reason a rule needs to be specific to a particular version of
WDL, it can implement the `document` visitor method and check the version of
the document; if it's not the appropriate version, the rule can disable itself.