refactor: make `Rule` a supertrait of `Visitor`.
Originally, `Visitor` was version specific, so the idea was that you would ask
a rule for its "V1" visitor, "V2" visitor, etc. to add to the validator.

However, that is no longer the case as `Visitor` is version agnostic.

Therefore, this commit removes the unnecessary "visitor" types implementing the
rules by having `Rule` be a supertrait of `Visitor`; now the rules themselves
implement the `Visitor` trait.

I also renamed `ExceptVisitor` to simply `LintVisitor` to make the API a little
cleaner; the visitor now means "run all lint rules and respect `except`
comments. The name `ExceptVisitor` was too specific.