feat: implement support for `requirements` sections in WDL 1.2
This commit implements support for the `requirements` section in WDL 1.2.

As the requirements keys are well-defined and the section does not accept
unsupported keys, a validation rule was added to ensure there are no unsupported
keys.

Additionally, conflicts between having both a runtime section and a
requirements section are checked for.

The `wdl-lint/src/v1.rs` file was deleted as it should have been removed by a
previous PR.