fix: make call target grammar rule more permissible.
Previously, the call target grammar rule accepted at most two identifiers as
part of the call target.

While this was technically to specification, if users added a third identifier
as part of the call target, it led to confusing error messages as recovery in
the middle of the call target would restart at the third identifier.

To address this, this changes the call target grammar rule from `<ident> ("."
<ident>)*` to `<ident> ("." <ident>)+`.

When analysis implements type checking (which will necessitate resolving call
statements), this will be treated as an error as we can currently only resolve
at most one namespace scope to find a task or workflow.