fix: remove the `Send` and `Sync` requirement on `Visitor`.
This removes the restriction on `Visitor` to be `Send` and `Sync`.

Originally the restriction was put in place in the hopes that they would be
needed to implement `wdl-analysis`; however, that turned out to not be the
case.

The restriction prevents visitors from storing AST elements as they are not
`Send` (only green nodes and tokens are `Send` and `Sync`).

Thus, rules would previously have to clone token string representations if
they were needed.

With these changes comes a `TokenStrHash` helper that can be used to hash an
AST token by its string representation; a few rules were updated to use it to
prevent unnecessary string clones.

Fixes #122.