feat: introduce `wdl-lsp` crate.
This commit introduces the `wdl-lsp` crate.

The new crate is responsible for implementing an LSP server that can be used in
compatible code editors.

The LSP server currently implements the following:

* Document diagnostics.
* Workspace diagnostics.
* Incremental document changes.
* Monitoring the workspace for file/directory changes.
* Support for multiple workspace folders.
* Notifying the client regarding work progress for long-running tasks.

To assist the implementation of the LSP, much if `wdl-analysis` was rewritten.

`AnalysisEngine` was renamed to simply `Analyzer` and much of its
implementation was moved to an internal "analysis queue" type.

The progress callback to the analyzer was made async so that an LSP server may
notify the client regarding the progress.