fix: allow reuse for `LintVisitor`.
Currently `LintVisitor` will remove rules from the rule set upon document entry
when there's a "global" exception.

This works for a single document visit, but if the visitor is reused, the
subsequent visitation will treat any previously globally excepted rules as
unknown rules.

The fix is to use the `each_enabled_rule` for the `document` method so that
globally excepted rules behave like locally excepted rules and the rule set
doesn't change.

Fixes #145.