fix: fix handling of `Union` arguments.
For type checking, an error in an expression might cause the result of the
expression to be treated as `Union` to convey the type is indeterminate.

Therefore, binding arguments to a function should handle accepting `Union`
arguments and also handle returning `Union` if, due to `Union` arguments, a
type parameter could not be inferred.

This also fixes `assert_valid` for `Type` to check that a compound type
definition identifier comes from the provided type collection.