feat: improve recovery for missing braces/brackets/parens.
This commit improves recovery around parsing of delimited items that end with a
closing brace, bracket, or paren.

The parser now maintains a stack of recovery token sets. When an item in a
delimited context fails to parse, we check to see if the token that starts the
failing item is also a recovery token in the immediately preceding context. If
so, we truncate the list of events at the failed node and reparse from the
parent context.

If we can't locate the closing brace/bracket/parens, we now synthesize a
closing token event with a zero-width token; this provides consistency in the
CST where matching opening/closing tokens are expected.

Also cleaned up the `ParserToken` trait a little bit.