fix: require commas for `input`, `output`, and `hints` literals.
This commit adds required commas to delineate items in `input`, `output`, and
`hints` literals.

These literals are modeled after literal objects, which have comma delimiters.

This was missed as the spec only has one example of more than one item in these
literals (introduced in 1.2), but that example does have a comma.