fix: ensure type checking does not occur on duplicate tasks.
Currently we limit the scope information based on task name, which will enable
us to quickly answer questions like "give me all the inputs of task `foo`" or
"what names are visible inside of task `foo`'s output section?".

However, if the document contains a duplicate task of the same name, we were
incorrectly using the first task's scope to evaluate the second task.

This change makes it so that we don't type check the duplicate task as we never
recorded a scope for it; instead, we rely on validation to output the
"duplicate task" diagnostic for the task.

This also changes the validation around duplicate tasks/workflows to ignore
the contained sections in the duplicates.