feat: complete type checking of tasks.
This commit completes the type checking in tasks to include expressions in
`requirements`, `hints`, and `runtime` sections.

Type checking works off of the "known" keys of those sections and the types
each key supports for an acceptable value.

For the `hints` section, this also includes supporting the hidden `input`,
`output`, and `hints` types and their literal expressions. For `input` and
`output` types, this means ensuring the input and output keys match a declared
input or output and also resolving any dotted form of the names when inputs and
outputs are structs.

Also fixed with these changes:

* Fixed a bug where duplicate tasks were generating duplicate section
  diagnostics instead of being ignored.
* Fixed a bug where duplicate tasks were being type checked with the wrong
  scope instead of being ignored.
* Fixed a failure to detect duplicate aliased items in task `hints` section.

Closes #166 and #164.