fix: parse workspace `hints` sections such that expressions are not supported.
This commit fixes the parsing of workflow `hints` sections such that the values
in the `hints` sections are always literals (no expressions supported).

To fix this, we need to parse the `workflow` hints section much like a metadata
section where expressions aren't supported.

It required splitting out the previous `HintsSection` AST representation into
two types: `TaskHintsSection` and `WorkflowHintsSection`, where the items in
the former support expressions for values and the latter does not.

Fixes #169.