refactor: merge task type checking with task scope creation.
This commit merges the type checking for tasks in with the creation of the task
within the document scope.

To do this, two passes are done over the AST's top level items: first imports
and struct definitions then tasks and workflow.

This allows the population of a task or a workflow in the document scope to
calculate type declarations up front and do the type checking when processing
the declaration, as declarations are added to the scope in topological order
now.