feat: implement static analysis for workflows.
This commit implements full static analysis for workflows, including type
checking.

To perform analysis on workflows, a `WorkflowEvaluationGraph` was implemented
that will return a topologically-sorted set of nodes for evaluating the
contents of a workflow in the correct order.

This also includes several fixes required to remove incorrect diagnostics from
checking the `workflows` repository:

* Treat a coercion to `T?` for a function argument of type `T` as a preference
  over any other coercion.
* Fix the signature of `select_any` such that it is monomorphic.
* Only consider signatures in overload resolution that have sufficient
  arguments.
* Allow coercion from `File` and `Directory` to `String`.
* Allow non-empty array literals to coerce to either empty or non-empty.
* Fix element type calculations for `Array` and `Map` so that `[a, b]`
  successfully calculates when `a` is coercible to `b`.
* Fix `if` expression type calculation such that `if (x) then a else b` works
  when `a` is coercible to `b`.
* Ensure that only equality/inequality expressions are supported on `File` and
  `Directory` now that there is a coercion to `String`.
* Allow index expressions on `Map`.

Included with these changes is a refactoring to place all of analysis
diagnostics in `diagnostics.rs`.

Fixes #197.
Fixes #196.
Fixes #167.
Fixes stjude-rust-labs/sprocket#21.