feat: add a `wdl-engine` crate.
This commit introduces a `wdl-engine` crate that will eventually be responsible
for implementing a WDL execution engine.

Currently the crate only defines two important types: `Engine` and `Value`.

`Engine` will be used to evaluate WDL source files, but is currently limited to
instantiating values.

`Value` represents a runtime WDL value; currently it implements the coercion
rules and a `display` method to format the value.

The next step is for `Engine` to analyze a WDL document and proceed to
evaluation if there are no diagnostics.

Once that is implemented, support for the WDL JSON format will be added as a
way to pass initial values into the evaluation.