chore: code review feedback.
Implement `rules` in `wdl-analysis` that has a representation of analysis
rules; this can be used by `sprocket explain` for listing available rules and
for explaining the rule.

Unlike `Rule` from `wdl-lint`, the `Rule` only stores the severity level for
the diagnostic; the rules themselves are implemented as part of analysis.

Also included is a unit test to ensure the reserved rule identifiers in
`wdl-lint` match those from `wdl-analysis`.