fix: normalize LSP request URIs.
This fixes LSP requests that contain percent encoded colons and
lowercase drive letters in request URIs.

This caused the analyzer to report an empty set of diagnostics because
the requested document was unknown to the analyzer (it used both a colon
and an uppercase drive letter in the document graph).

The fix is to percent decode the path segments and normalize the drive
letter to uppercase in the LSP server.

Fixes #229.