refactor: refactor the analysis API.
This commit refactors the analysis API to assist in future engine work.

Primarily it exposes `Task` and `Workflow` from analyzed documents so that a
task or workflow's inputs and outputs can be easily determined.

Additionally, this renames `DocumentScope` to simply `Document` to represent an
analyzed document.

Scopes are now kept internally to a `Task` or `Workflow`, which allows for a
`scope` method directly on those types.

As a result of the rename, `scope.rs` has been renamed to `document.rs`.