refactor: refactor call types in `wdl-analysis`.
This commit refactors `CallOutputType` to `CallType` so that the call can
represent both the calls inputs and outputs in the type information.

Additionally, workflows now expose a map of call names to call types.

This allows the analyzed document to describe the calls a workflow makes so
that nested inputs can be validated.

Also fixes a bug where the local call name was being used in diagnostics
instead of the original task or workflow name.