refactor: return more information from `Function::bind`.
Previously, only the calculated return type was returned from `Function::bind`.

This change adds both the overload index and the bound function signature are
returned as well.

The index can then be used to statically dispatch the call in expression
evaluation.

Additionally, the minimum version has moved from being specified on the
function to each individual signature.

This will allow the standard to add overloads for existing functions in the
future.