feat: implement the `write_tsv` function.
This commit implements the `write_tsv` function from the WDL standard library.

Additionally, it fixes some overloads of `read_tsv` and `write_tsv` that should
be marked as minimum version 1.2.

The constraints on the file functions that take arrays of structs should also
be further constrained on the requirement that the structs only contain
primitive members.