feat: implement workflow evaluation.
This commit implements workflow evaluation in `wdl-engine` and in the `wdl run`
command.

Workflow evaluation works by taking the workflow evaluation graph and
"splitting" the graph into subgraphs at each entry into a conditional and
scatter statement.

For conditional statements, their subgraphs are only evaluated if the
conditional expression evaluates to true.

For scatter statements, their subgraphs are evaluated for each element in the
scatter array.

Workflow evaluation is currently single-threaded. Progress is made on the
evaluation graph as much as possible until no more progress can be made, at
which point evaluation will wait for outstanding tasks to complete.

Currently local task execution is limited to the amount of available
parallelism supported on the host system; in the future, we'll want that to be
configurable.