refactor: make `Analyzer` accept the diagnostics configuration.
This commit refactors the `new` and `new_with_validator` methods of `Analyzer`
to directly accept the `DiagnosticsConfig` rather than an iterator of rules
which it immediately uses to create a `DiagnosticsConfig` internally.

By having the `DiagnosticsConfig` (which is `Copy`) passed in directly, we
don't have to borrow a rule exception list that might also be used as part of
the validator closure passed to `new_with_validator`, sparing us a clone.