refactor: reduce size of the `Value` type.
Currently, the `Task` variant of the `Value` enum causes `Value` to be about
200 bytes in size due to its numerous fields.

This change puts the immutable fields of a task value behind an indirection,
reducing the size of `Value` to 40 bytes.

It also puts the `left` and `right` members of `Pair` behind a single
allocation rather than two.