feat: implement the `env` declaration modifier in WDL 1.2.
This commit implements the `env` declaration modifier in WDL 1.2 that may be
present for both inputs and private declarations in a task.

When present, task execution will treat the value of the declaration as an
environment variable for the spawned process. The variable can then be
referenced using Bash syntax (e.g. `$foo`).

Validation diagnostics were added to ensure the `env` variable is only applied
to primitive type declarations, as there is no standard string representation
for complex types.

Closes #295.