fix: `wdl run` should join input paths correctly.
This commit fixes the `wdl run` command to properly join input paths relative
to the inputs file.

The code to make the path of the inputs file absolute was incorrectly done,
resulting in the later `join_paths` call to effectively be a no-op when the
inputs file path on the command line was relative to CWD (e.g. `./foo.json`).

As a result, task execution would likely fail if using the path as the path
would then be treated as relative to the task's working directory because it
was not made absolute.

The fix is to simply return the absolute path correctly from `parse_inputs`.

Fixes #301.