feat: report progress during `wdl run`.
This commit implements reporting progress during `wdl run`. The
`WorkflowEvaluator` now accepts a callback that is called when a task or
workflow is evaluated and when it is completed.

The progress bar displayed by `wdl run` has been updated to print out the
number of completed tasks, the number of in-process tasks, including up to 10
names of tasks currently being executed.