fix: use the correct type for scatter output.
This commit fixes an incorrect type used when creating the output of a scatter
statement into the parent scope.

It should be an array type, but the array's element type was being used
instead.

The fix is to use the proper type and assert, at a minimum, that the
`new_unchecked` functions at least get the expected compound type.

Fixes #315.