fix: fix retry logic in task evaluation.
This commit fixes the retry logic in task evaluation to re-evaluate the command
section before retrying.

The change enables a command to see an updated `task` variable in 1.2 and can
therefore know when it has failed in a previous attempt.

Additionally, this makes it so that we also re-evaluate the hints and
requirements sections upon retry; the hope is that in the near future the spec
will be updated to allow those sections to have task scope.

A test was added to retry a task.