feat: implement mount points for task execution.
This commit implements the calculation of mount points for task execution
backends that use containers.

Paths from `File` and `Directory` values that are visible prior to evaluating
the command section of a task are added to a `PathTrie`.

The `PathTrie` is used to determine the minimum number of mounts in the guest
environment to be able to send input files to the container.

Command evaluation does a translation from host paths to guest paths so that
the generated command script uses only guest paths.

Similarly, evaluation of output values does a translation from guest paths to
host paths so that tasks always output the host path.

Fixes #267.