feat: implement support for a Crankshaft backend.
This commit implements support for using Crankshaft as a task
execution backend.

Currently the Crankshaft backends are limited to Docker, but the
plan is to expand the support out to all of Crankshafts supported
backends in the future.

Note that this takes a git dependency on the `crankshaft` crate as
it has not yet been published; additionally, a patch to `bollard`
is necessary which hasn't been upstreamed yet (but will be soon).

As a result, we will not be able to release `wdl` until the next
release of `bollard` that includes the patch.

The local backend has been refactored to share resource management
code with the Crankshaft backend when it is configured to use
the Docker backend that is connected to a Docker daemon that is
not in a swarm.

Calculation of mounts has changed so we only map relative paths to
a mounted directory for paths in the trie that are children of a
terminal node; this allows for the Crankshaft backend to receive
information about the task inputs in a format it expects. The way
it is currently done minimizes for the mounts, but doesn't convey
enough information for backends like TES that may need to
upload individual input files.

The `cpu` and `memory` requirements and the `max_cpu` and `max_memory`
hints are also implemented for the Docker backend.

Additionally, more context is now provided for tasks that fail, such
as the name of the task and the path to the file containing the task.

The diagnostic returned from a failed call has also now been corrected
to point at the call statement instead of an incorrect span when
the call is to an imported task or workflow.

Cancellation has been implemented so that backend resources are
cleaned up; note that the "failure mode" is still a fast one, meaning
we do not wait for currently executing containers to exit, we end up
sending them a SIGKILL signal instead. In the future, we'll want a
cancellation mode that will wait for the containers to exit to update
the call cache so that successful task executions don't need to run
again.

Added support for engine configuration via a TOML file to `wdl run`.

Fixes #319.